move CloseRef into chorus subpackage
